ethclient/lightclient: split up chain.go and reordered functions
